refactored test suite to autogenerate data. might want to look into factory-boy
