+++ OpenOversight/tests/conftest.py
@@ -1,10 +1,70 @@
 import os
 import pytest
-
+from flask import current_app
 from app import create_app
 from app import db as _db
 from app import models
 from datetime import datetime
+import random
+
+OFFICERS = [('IVANA', '', 'TINKLE'),
+            ('SEYMOUR', '', 'BUTZ'),
+            ('HAYWOOD', 'U', 'CUDDLEME'),
+            ('BEA', '', 'O\'PROBLEM'),
+            ('URA', '', 'SNOTBALL'),
+            ('HUGH', '', 'JASS')]
+
+def pick_birth_date():
+    return random.randint(1950, 2000)
+
+
+def pick_race():
+    return random.choice(['WHITE', 'BLACK', 'HISPANIC', 'ASIAN',
+                         'PACIFIC ISLANDER'])
+
+def pick_gender():
+    return random.choice(['M', 'F'])
+
+def pick_first():
+    return random.choice(OFFICERS)[0]
+
+def pick_middle():
+    return random.choice(OFFICERS)[1]
+
+def pick_last():
+    return random.choice(OFFICERS)[2]
+
+def pick_name():
+    return (pick_first(), pick_middle(), pick_last())
+
+def pick_rank():
+    return random.choice(['COMMANDER', 'CAPTAIN', 'PO'])
+
+def pick_star():
+    return random.randint(1, 9999)
+
+def generate_officer():
+    year_born = pick_birth_date()
+    f_name, m_initial, l_name = pick_name()
+    return models.Officer(
+        last_name=l_name, first_name=f_name,
+        middle_initial=m_initial,
+        race=pick_race(), gender=pick_gender(),
+        birth_year=year_born,
+        employment_date=datetime(year_born + 20, 4, 4, 1, 1, 1),
+        pd_id=1
+    )
+
+def build_assignment(officer):
+    return models.Assignment(star_no=pick_star(),
+                      rank=pick_rank(),
+                      officer=officer)
+
+def assign_faces(officer, images):
+    if random.uniform(0, 1) >= 0.5:
+        return models.Face(officer_id=officer.id, img_id=random.choice(images).id)
+    else:
+        return False
 
 @pytest.fixture(scope='session')
 def app(request):
@@ -16,7 +76,7 @@ def app(request):
     ctx.push()
 
     def teardown():
-	ctx.pop()
+        ctx.pop()
 
     request.addfinalizer(teardown)
     return app
@@ -28,8 +88,8 @@ def db(app, request):
     #     os.unlink(TESTDB_PATH)
 
     def teardown():
-	_db.drop_all()
-	# os.unlink(TESTDB_PATH)
+        _db.drop_all()
+        # os.unlink(TESTDB_PATH)
 
     _db.app = app
     _db.create_all()
@@ -49,45 +109,18 @@ def session(db, request):
     db.session = session
 
     def teardown():
-	transaction.rollback()
-	connection.close()
-	session.remove()
+        transaction.rollback()
+        connection.close()
+        session.remove()
 
     request.addfinalizer(teardown)
     return session
 
 @pytest.fixture
 def mockdata(session, request):
-    po1 = models.Officer(last_name='TINKLE', first_name='IVANA', race='BLACK',
-			 gender='F', employment_date=datetime(2000, 4, 4, 1, 1, 1),
-			 birth_year=1970, pd_id=1)
-    po2 = models.Officer(last_name='JASS', first_name='HUGH', race='WHITE',
-			 gender='M', birth_year=1950, pd_id=1,
-			 employment_date=datetime(1996, 4, 4, 1, 1, 1))
-    po3 = models.Officer(last_name='Butz', first_name='Seymour', race='WHITE',
-			 gender='F', birth_year=1950, pd_id=1,
-			 employment_date=datetime(1983, 4, 4, 1, 1, 1))
-    po4 = models.Officer(last_name='CUDDLEME', first_name='HAYWOOD', middle_initial='U',
-			 race='HISPANIC', gender='F', birth_year=1950, pd_id=1,
-			 employment_date=datetime(2014, 4, 4, 1, 1, 1))
-    po5 = models.Officer(last_name='KLOZOFF', first_name='OLIVER', middle_initial='U',
-			 race='WHITE', gender='M', birth_year=1950, pd_id=1,
-			 employment_date=datetime(2004, 4, 4, 1, 1, 1))
-    po6 = models.Officer(last_name='O\'PROBLEM', first_name='BEA', middle_initial='U',
-			 race='HISPANIC', gender='F', birth_year=1978, pd_id=1,
-			 employment_date=datetime(2014, 4, 4, 1, 1, 1))
-
-    test_officers = [po1, po2, po3, po4, po5, po6]
-    session.add_all(test_officers)
-
-    star1 = models.Assignment(star_no=1234, rank='COMMANDER', officer=po1)
-    star2 = models.Assignment(star_no=5678, rank='PO', officer=po2)
-    star3 = models.Assignment(star_no=9012, rank='CHIEF', officer=po3)
-    star4 = models.Assignment(star_no=3456, rank='LIEUTENANT', officer=po4)
-    star5 = models.Assignment(star_no=5227, rank='PO', officer=po5)
-    star6 = models.Assignment(star_no=9120, rank='DEPUTY CHIEF', officer=po6)
-    test_assignments = [star1, star2, star3, star4, star5, star6]
-    session.add_all(test_assignments)
+    NUM_OFFICERS = current_app.config['NUM_OFFICERS']
+    SEED = current_app.config['SEED']
+    random.seed(SEED)
 
     image1 = models.Image(filepath='static/images/test_cop1.png')
     image2 = models.Image(filepath='static/images/test_cop2.png')
@@ -95,23 +128,21 @@ def mockdata(session, request):
     image4 = models.Image(filepath='static/images/test_cop4.png')
 
     test_images = [image1, image2, image3, image4]
+    officers = [generate_officer() for o in range(NUM_OFFICERS)]
+    assignments = [build_assignment(officer) for officer in officers]
+    faces = [assign_faces(officer, test_images) for officer in officers]
+    faces = [f for f in faces if f]
     session.add_all(test_images)
-
-    face1 = models.Face(officer_id=po1.id, img_id=image1.id)
-    face2 = models.Face(officer_id=po2.id, img_id=image2.id)
-    face3 = models.Face(officer_id=po3.id, img_id=image3.id)
-    face4 = models.Face(officer_id=po4.id, img_id=image4.id)
-    face5 = models.Face(officer_id=po1.id, img_id=image3.id)
-
-    test_faces = [face1, face2, face3, face4, face5]
-    session.add_all(test_faces)
-
+    session.add_all(officers)
+    session.add_all(assignments)
+    session.add_all(faces)
     session.commit()
+    return assignments[0].star_no
 
 @pytest.fixture
 def client(app, request):
     client = app.test_client()
     def teardown():
-	pass
+        pass
     request.addfinalizer(teardown)
     return client
